Rename file to match contained class
