+++ lib/EntryListManager.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'storage_manager.dart';
-import 'entry.dart';
-import 'filterOptions.dart';
+import 'StorageManager.dart';
+import 'Entry.dart';
+import 'FilterOptions.dart';
 
 class EntryListManager with ChangeNotifier{
   List<Entry> globalList = <Entry>[];
